+++ scrapers/nd/events.py
@@ -1,114 +1,174 @@
-import dateutil.parser
-import lxml
-import pytz
 import re
+import pytz
+import logging
+import dateutil.parser
+import datetime
+import requests
+import lxml.html
+from spatula import HtmlPage
 from openstates.scrape import Scraper, Event
 
 
-class NDEventScraper(Scraper):
-    _tz = pytz.timezone("US/Central")
-    event_months = set()
-    events = {}
+def time_is_earlier(new, current):
+    new_magnitude = int(new[0:2] + new[3:5])
+    current_magnitude = int(current[0:2] + current[3:5])
+    if new_magnitude < current_magnitude:
+        return True
+    else:
+        return False
 
-    def scrape(self, session=None):
 
-        if not session:
-            session = self.latest_session()
-            self.info("no session specified, using %s", session)
-        # figuring out starting year from metadata
-        for item in self.jurisdiction.legislative_sessions:
-            if item["identifier"] == session:
-                start_year = item["start_date"][:4]
-                self.year = start_year
-                break
+class EventConsolidator:
+    def __init__(self, items, url):
+        self.items = items
+        self.events = {}
+        self._tz = pytz.timezone("US/Central")
+        self.url = url
 
-        url = f"https://www.legis.nd.gov/assembly/{session}-{start_year}/committees/interim/committee-meeting-summary"
+    def consolidate(self):
+        for item in self.items:
+            date_time = item["date_time"]
+            local_date = dateutil.parser.parse(date_time)
 
-        page = self.get(url).content
-        page = lxml.html.fromstring(page)
-        page.make_links_absolute(url)
+            item_date, item_time = str(local_date).split()
 
-        for table in page.xpath('//table[contains(@class,"views-table")]'):
-            com = table.xpath("caption/a")[0].text_content().strip()
-            for row in table.xpath("tbody/tr"):
-                date_link = row.xpath("td[1]/strong/a")[0]
-                event_url = date_link.xpath("@href")[0]
+            committee = item["committee"]
+            location = item["location"]
+            event_key = f"{item_date}-{committee}-{location}"
 
-                date = date_link.xpath("span")[0].text_content().strip()
-                date = dateutil.parser.parse(date)
-                date = self._tz.localize(date)
+            if not self.events.get(event_key):
+                self.events[event_key] = {}
+                self.events[event_key]["event_start_time"] = item_time
+                self.events[event_key]["item_keys"] = set()
+            else:
+                current_start = self.events[event_key]["event_start_time"]
+                if time_is_earlier(item_time, current_start):
+                    self.events[event_key]["event_start_time"] = item_time
 
-                self.event_months.add(date.strftime("%Y-%m"))
+            self.events[event_key]["item_keys"].add(item_time)
 
-                location = "See Agenda"
+            agenda_item_details = {
+                "description": item["description"],
+                "bill_name": item["bill_name"],
+                "sub_com": item["sub_com"],
+            }
+            self.events[event_key][item_time] = []
+            self.events[event_key][item_time].append(agenda_item_details)
 
-                event = Event(name=com, start_date=date, location_name=location)
+        yield from self.create_events()
 
-                event.add_source(event_url)
+    def create_events(self):
+        event_collection = []
 
-                for link in row.xpath("td[2]//a"):
-                    link_text = link.text_content().strip()
+        for event in self.events.keys():
+            date, com, loc = re.search(r"(.+)\-(.+)\-(.+)", event).groups()
 
-                    # skip live broadcast links
-                    if "video.legis" in link_text:
-                        continue
+            start_time = self.events[event]["event_start_time"]
+            date_time = f"{date} {start_time}"
+            date_object = dateutil.parser.parse(date_time)
+            date_with_offset = self._tz.localize(date_object)
 
-                    event.add_document(
-                        link_text, link.xpath("@href")[0], media_type="application/pdf"
-                    )
-
-                self.events[event_url] = event
-
-        for year_month in self.event_months:
-            self.scrape_calendar(year_month)
-
-        for key in self.events:
-            yield self.events[key]
-
-    # the listing page has all the events and attachments, but no dates or locations
-    # the calendar page has dates and locations, but no attachments
-    def scrape_calendar(self, year_month):
-        self.info(f"Scraping calendar for {year_month}")
-        cal_url = f"https://ndlegis.gov/events/{year_month}"
-        page = self.get(cal_url).content
-        page = lxml.html.fromstring(page)
-        page.make_links_absolute(cal_url)
-
-        for row in page.xpath("//div[contains(@class,'calendar-item')]"):
-            if row.xpath(
-                ".//div[contains(@class,'event-count') and (contains(text(), '2 of') or contains(text(), '3 of'))]"
-            ):
-                self.info("Skipping later day of multi-day event")
-                continue
-
-            if not row.xpath(".//span[contains(@class,'date-display-start')]"):
-                continue
-
-            event_url = row.xpath(".//div[contains(@class,'event-title')]/a/@href")[0]
-
-            if event_url not in self.events:
-                self.info(f"Skipping {event_url}, not on hearing listing page.")
-                continue
-
-            event_time = row.xpath(
-                ".//span[contains(@class,'date-display-start')]/text()"
-            )[0]
-            event_date = (
-                self.events[event_url].as_dict()["start_date"].strftime("%Y-%m-%d")
+            event_obj = Event(
+                name=com,
+                location_name=loc,
+                description="Standing Committee Hearing",
+                start_date=date_with_offset,
             )
-            new_start = f"{event_date} {event_time}"
 
-            date = dateutil.parser.parse(new_start)
-            date = self._tz.localize(date)
-            self.events[event_url].__setattr__("start_date", date)
+            for item_key in self.events[event]["item_keys"]:
+                agenda = self.events[event][item_key]
+                for item in agenda:
+                    time = datetime.datetime.strptime(item_key, "%H:%M:%S").strftime(
+                        "%I:%M %p"
+                    )
+                    descr_with_time = f"[{time}]: {item['description']}"
+                    item_descr = event_obj.add_agenda_item(descr_with_time)
+                    if item["bill_name"]:
+                        item_descr.add_bill(item["bill_name"])
+                    if item["sub_com"]:
+                        item_descr.extras["sub_committee"] = item["sub_com"]
 
-            if row.xpath(".//div[contains(@class,'vcard')]"):
-                raw_loc = row.xpath(".//div[contains(@class,'vcard')]")[
-                    0
-                ].text_content()
-                loc = raw_loc.strip()
-                loc = re.sub(r"\s+", " ", loc)
-                loc = loc.replace("State Capitol", "600 East Boulevard Avenue")
-                loc_dict = {"name": loc, "note": "", "coordinates": None}
-                if len(loc) > 0:
-                    self.events[event_url].__setattr__("location", loc_dict)
+            event_obj.add_source(self.url)
+
+            event_collection.append(event_obj)
+
+        return event_collection
+
+
+class BillNameScraper(HtmlPage):
+    def __init__(self, source):
+        self.source = source
+
+    def get_bill_name(self):
+        response = requests.get(self.source)
+        content = lxml.html.fromstring(response.content)
+        bill_name_tag = content.xpath(".//div[@id='content']//h3")[0]
+        bill_name = bill_name_tag.text_content().strip()
+        return bill_name
+
+
+class EventsTable(HtmlPage):
+    events_path = "/legend/committee/hearings/public-schedule/"
+    source = f"https://www.ndlegis.gov{events_path}"
+
+    def process_page(self):
+        date_span = self.root.xpath(".//h5[2]")[0].text_content()
+        start_year, end_year = re.search(r"(\d{4}).+(\d{4})", date_span).groups()
+
+        table_rows = self.root.xpath(".//tbody//tr")
+        agenda_items_list = []
+
+        for row_item in table_rows:
+            columns = row_item.xpath("./td")
+            columns_content = [x.text_content().strip() for x in columns]
+            if len(columns_content) == 7:
+                columns_content = columns_content[:-1]
+
+            bill, part_date, com, sub_com, loc, descr = columns_content
+
+            full_bill_name_match = re.search(r"[A-Z]{1,4}\s+\d+", bill)
+            if full_bill_name_match:
+                bill_name = full_bill_name_match.group()
+            else:
+                partial_bill_name_match = re.search(r"\d+", bill)
+                if partial_bill_name_match:
+                    match_in_descr = re.search(r"[A-Z]{1,4}\s{0,3}\d+", descr)
+                    if match_in_descr:
+                        bill_name = match_in_descr.group()
+                    else:
+                        bill_link = columns[0].xpath("./a")[0].get("href")
+                        bill_name_scraper = BillNameScraper(bill_link)
+                        bill_name = bill_name_scraper.get_bill_name()
+                else:
+                    bill_name = ""
+
+            date_time_parts = part_date.split()
+            month = int(re.search(r"\d+", date_time_parts[0]).group())
+            if month == 1 and not start_year == end_year:
+                event_year = end_year
+            else:
+                event_year = start_year
+            date_time_parts[0] = date_time_parts[0] + f"/{event_year}"
+            date_time = " ".join(date_time_parts)
+
+            agenda_item = {
+                "bill_name": bill_name,
+                "date_time": date_time,
+                "committee": com,
+                "sub_com": sub_com,
+                "location": loc,
+                "description": descr,
+            }
+
+            agenda_items_list.append(agenda_item)
+
+        events = EventConsolidator(agenda_items_list, self.source.url)
+        return events.consolidate()
+
+
+class NDEventScraper(Scraper):
+    @staticmethod
+    def scrape():
+        logging.getLogger("scrapelib").setLevel(logging.WARNING)
+        event_list = EventsTable()
+        yield from event_list.do_scrape()
